Remove nocodb-daily.js and package-lock.json
Files 'nocodb-daily.js' and 'package-lock.json' were deleted as these were not needed. The 'nocodb-daily.js' was an experimental trial that is not a part of the final project. In addition, 'package-lock.json' was removed as npm instability issues were causing unnecessary dense diffs. As the project has a specific 'package.json', there's no need to keep 'package-lock.json' version controlled.